refactor: update CuratorsRoute and CuratorsClient components to use children prop for ComponentAuthGuard, improving readability and consistency in component structure
