+++ packages/solito/apps/expo/app/profile/curators.tsx
@@ -1,7 +1,11 @@
+import { ComponentAuthGuard } from 'app/utils/authGuard'
 import CuratorsScreen from 'app/features/profile/curators'
-import { ComponentAuthGuard } from 'app/features/auth/authGuard'
 
 export default function CuratorsRoute() {
-  return <ComponentAuthGuard component={<CuratorsScreen />} />
+  return (
+    <ComponentAuthGuard>
+      <CuratorsScreen />
+    </ComponentAuthGuard>
+  )
 }
